# 8299807: newStringNoRepl should avoid copying arrays for ASCII compatible charsets