This is the javadoc of `JavaLangAccess::newStringNoRepl`:

```java
    /**
     * Constructs a new {@code String} by decoding the specified subarray of
     * bytes using the specified {@linkplain java.nio.charset.Charset charset}.
     *
     * The caller of this method shall relinquish and transfer the ownership of
     * the byte array to the callee since the later will not make a copy.
     *
     * @param bytes the byte array source
     * @param cs the Charset
     * @return the newly created string
     * @throws CharacterCodingException for malformed or unmappable bytes
     */
```

It is recorded in the document that it should be able to directly construct strings with parameter byte array to reduce array allocation.

However, at present, `newStringNoRepl` always copies arrays for UTF-8 or other ASCII compatible charsets.

This PR fixes this problem.

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must be properly reviewed (1 review required, with at least 1 [Reviewer](https://openjdk.org/bylaws#reviewer))
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue

### Issue
 * [JDK-8299807](https://bugs.openjdk.org/browse/JDK-8299807): newStringNoRepl should avoid copying arrays for ASCII compatible charsets


### Reviewers
 * [Roger Riggs](https://openjdk.org/census#rriggs) (@RogerRiggs - **Reviewer**)


### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.org/jdk pull/12119/head:pull/12119` \
`$ git checkout pull/12119`

Update a local copy of the PR: \
`$ git checkout pull/12119` \
`$ git pull https://git.openjdk.org/jdk pull/12119/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 12119`

View PR using the GUI difftool: \
`$ git pr show -t 12119`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.org/jdk/pull/12119.diff">https://git.openjdk.org/jdk/pull/12119.diff</a>

</details>