@Glavo 2023-01-20
Can someone help me open an Issue on Java Bug System?