@Glavo 2023-01-21
The throughput of reading ASCII files as UTF-8:

![image](https://user-images.githubusercontent.com/20694662/213833461-97e60b8d-2845-48fb-9331-585810c182b2.png)


The throughput of reading ASCII files as GBK:

![image](https://user-images.githubusercontent.com/20694662/213833687-d2811afb-efce-4d4e-9074-a797f2eea83a.png)