@RogerRiggs 2023-01-27
It seems odd that the benchmark seems slower for smaller files; can you suggest why that might be?
I'd expect the size distribution for Files.readString to be biased toward the smaller files.
Can you repeat the benchmark using the default file system.  OS file caching should eliminate the disk speed effects.