@Glavo 2023-01-30
Benchmark results for large files (128KiB~256MiB) on the default file system:

![image](https://user-images.githubusercontent.com/20694662/215372915-2988bb4e-8c35-4def-9fe9-a6bc4f033708.png)

![image](https://user-images.githubusercontent.com/20694662/215372925-2a7995dc-00f4-4a91-9b2a-e73c97192008.png)

For files of about 1MiB, the throughput can be improved by more than 55%.

Original results: https://gist.github.com/Glavo/f3d2060d0bd13cd0ce2add70e6060ea0?permalink_comment_id=4452798#gistcomment-4452798